+++ src/components/loadings/Loading.vue
@@ -1,21 +1,11 @@
 <template>
     <div class="inline-flex relative w-[60px] h-[60px]">
         <div v-for="i in 4" :key="i"
-            class="box-border absolute w-4/5 h-4/5 m-2 border-[8px] border-transparent rounded-full animate-ring"
-            :class="[`animate-delay-${(i - 1) * 150}`]" :style="{
-                borderTopColor: themeClasses.hex,
-            }">
+            class="box-border absolute w-4/5 h-4/5 m-2 border-[8px] border-transparent rounded-full animate-ring border-t-vdt-primary-500"
+            :class="[`animate-delay-${(i - 1) * 150}`]">
         </div>
     </div>
 </template>
-
-<script setup lang="ts">
-import { inject, type ComputedRef } from 'vue';
-import type { ThemeStateClasses } from '@/types/internal';
-
-const themeClasses = inject<ComputedRef<ThemeStateClasses>>('themeClasses')!;
-
-</script>
 
 <style scoped>
 @keyframes ring {