+++ src/components/loadings/LoadingLine.vue
@@ -1,15 +1,8 @@
 <template>
     <div class="w-full h-[3px] relative overflow-hidden bg-gray-300">
-        <div class="absolute h-[3px] w-2/5 animate-loading-line" :style="{ backgroundColor: themeClasses.hex }"></div>
+        <div class="absolute h-[3px] w-2/5 animate-loading-line bg-vdt-primary-500"></div>
     </div>
 </template>
-
-<script setup lang="ts">
-import { inject, type ComputedRef } from 'vue';
-import type { ThemeStateClasses } from '@/types/internal';
-
-const themeClasses = inject<ComputedRef<ThemeStateClasses>>('themeClasses')!;
-</script>
 
 <style scoped>
 @keyframes loading-line {